Добавлены новые страницы и компоненты для генерации и отображения гитарных цепочек. Реализованы API для генерации, получения и удаления цепочек, а также для обновления статуса генераций. Обновлены конфигурации TypeScript и Tailwind, добавлены метаданные для страниц. Удалены устаревшие файлы и компоненты.
