+++ .eslintrc.json
@@ -24,11 +24,11 @@
           { "argsIgnorePattern": "^_" }
         ],
         "@typescript-eslint/explicit-function-return-type": [
-          "warn",
+          "off",
           { "allowExpressions": false }
         ],
-        "@typescript-eslint/explicit-module-boundary-types": ["warn"],
-        "@typescript-eslint/require-await": ["warn"],
+        "@typescript-eslint/explicit-module-boundary-types": ["off"],
+        "@typescript-eslint/require-await": ["off"],
         "@typescript-eslint/typedef": [
           "error",
           {
