+++ components/PresetCard.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import { QRCodeCanvas } from "qrcode.react";
 import Image from "next/image";
 import Link from "next/link";
@@ -30,32 +32,36 @@ export function PresetCard({ preset }: PresetCardProps): React.ReactElement {
       </div>
 
       <div className="relative p-6">
-        {/* QR Code */}
-        <div className="flex justify-center mb-6">
-          <div className="bg-white p-2 rounded">
-            <QRCodeCanvas value={qrCode.qrCode} size={180} level="M" />
+        <div className="flex flex-col md:flex-row items-center gap-6">
+          {/* QR Code */}
+          <div className="shrink-0">
+            <div className="bg-white p-2 rounded">
+              <QRCodeCanvas value={qrCode.qrCode} size={180} level="M" />
+            </div>
           </div>
-        </div>
 
-        <div className="text-center">
-          <div className="text-gray-300">
-            <div className="text-lg font-medium">{preset.origin.song}</div>
-            <div className="text-sm text-gray-400">{preset.origin.artist}</div>
-            {preset.origin.part && (
-              <div className="text-xs text-gray-500 mt-1">
-                {preset.origin.part}
+          <div className="text-center md:text-left flex-grow">
+            <div className="text-gray-300">
+              <div className="text-lg font-medium">{preset.origin.song}</div>
+              <div className="text-sm text-gray-400">
+                {preset.origin.artist}
               </div>
-            )}
-          </div>
+              {preset.origin.part && (
+                <div className="text-xs text-gray-500 mt-1">
+                  {preset.origin.part}
+                </div>
+              )}
+            </div>
 
-          {/* Device and Instrument Type */}
-          <div className="mt-4 flex items-center justify-center gap-2 text-xs">
-            <span className="bg-pink-500/20 text-pink-400 px-2 py-1 rounded">
-              Guitar
-            </span>
-            <span className="bg-violet-500/20 text-violet-400 px-2 py-1 rounded">
-              NUX Mighty Plug Pro
-            </span>
+            {/* Device and Instrument Type */}
+            <div className="mt-4 flex items-center justify-center md:justify-start gap-2 text-xs">
+              <span className="bg-pink-500/20 text-pink-400 px-2 py-1 rounded">
+                Guitar
+              </span>
+              <span className="bg-violet-500/20 text-violet-400 px-2 py-1 rounded">
+                NUX Mighty Plug Pro
+              </span>
+            </div>
           </div>
         </div>
       </div>
